# swagger-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinanceBudget",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transactions": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Transaction",
        "operationId": "create_transaction_api_v1_transactions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_v1_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Sort"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTransactions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Delete Transaction",
        "operationId": "delete_transaction_api_v1_transactions__transaction_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/budgets": {
      "post": {
        "tags": [
          "budgets"
        ],
        "summary": "Create Budget",
        "description": "Set a monthly spending limit for a category.",
        "operationId": "create_budget_api_v1_budgets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "budgets"
        ],
        "summary": "List Budgets",
        "description": "List all budgets, optionally filtered by month and year.",
        "operationId": "list_budgets_api_v1_budgets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BudgetResponse"
                  },
                  "title": "Response List Budgets Api V1 Budgets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/budgets/status": {
      "get": {
        "tags": [
          "budgets"
        ],
        "summary": "Budget Status",
        "description": "Returns spending vs limit for every budget set in the given month.\nIncludes amount spent, remaining, percentage used, and a status flag.",
        "operationId": "budget_status_api_v1_budgets_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 2000,
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/budgets/{budget_id}": {
      "delete": {
        "tags": [
          "budgets"
        ],
        "summary": "Delete Budget",
        "description": "Remove a monthly budget limit.",
        "operationId": "delete_budget_api_v1_budgets__budget_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Budget Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BudgetCreate": {
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "month": {
            "type": "string",
            "format": "date",
            "title": "Month"
          },
          "limit_amount": {
            "type": "number",
            "title": "Limit Amount"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "month",
          "limit_amount"
        ],
        "title": "BudgetCreate"
      },
      "BudgetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "month": {
            "type": "string",
            "format": "date",
            "title": "Month"
          },
          "limit_amount": {
            "type": "number",
            "title": "Limit Amount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "month",
          "limit_amount"
        ],
        "title": "BudgetResponse"
      },
      "BudgetStatusItem": {
        "properties": {
          "budget_id": {
            "type": "string",
            "format": "uuid",
            "title": "Budget Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "limit_amount": {
            "type": "number",
            "title": "Limit Amount"
          },
          "spent": {
            "type": "number",
            "title": "Spent"
          },
          "remaining": {
            "type": "number",
            "title": "Remaining"
          },
          "percentage_used": {
            "type": "number",
            "title": "Percentage Used"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "budget_id",
          "category_id",
          "category_name",
          "limit_amount",
          "spent",
          "remaining",
          "percentage_used",
          "status"
        ],
        "title": "BudgetStatusItem"
      },
      "BudgetStatusResponse": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BudgetStatusItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_budget": {
            "type": "number",
            "title": "Total Budget"
          },
          "total_spent": {
            "type": "number",
            "title": "Total Spent"
          },
          "total_remaining": {
            "type": "number",
            "title": "Total Remaining"
          }
        },
        "type": "object",
        "required": [
          "month",
          "items",
          "total_budget",
          "total_spent",
          "total_remaining"
        ],
        "title": "BudgetStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedTransactions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedTransactions"
      },
...
      "TransactionCreate": {
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "category_id",
          "amount",
          "type",
          "date"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "category_id",
          "amount",
          "type",
          "date",
          "description"
        ],
        "title": "TransactionResponse"
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ],
        "title": "TransactionType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}